Remove functors.hpp.
